# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation Netpro",
        "description": "Dokumentasi API untuk Netpro",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/mikrotik/Check-Vpn": {
            "post": {
                "tags": [
                    "OpenVPN"
                ],
                "summary": "Cek apakah interface VPN sudah ada di Mikrotik",
                "description": "Memeriksa apakah interface dengan nama tertentu sudah ada pada Mikrotik",
                "operationId": "7be81440c5d09fdda74a99832dad7e53",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "interface_name"
                                ],
                                "properties": {
                                    "interface_name": {
                                        "type": "string",
                                        "example": "ovpn-client1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil mengecek interface",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "exists": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "VPN Sudah Ada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal terhubung ke Mikrotik",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Gagal terhubung ke Mikrotik"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connection refused"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "X-Tenant-ID": []
                    }
                ]
            }
        },
        "/configure-vpn-server": {
            "post": {
                "tags": [
                    "OpenVPN"
                ],
                "summary": "Konfigurasi VPN di Mikrotik",
                "description": "Membuat konfigurasi OpenVPN baru di Mikrotik",
                "operationId": "79adec3460cf866d44d0c077b2d0daa2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "pool_name",
                                    "port_Nat",
                                    "port_address"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "vpnuser"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "securepass"
                                    },
                                    "pool_name": {
                                        "type": "string",
                                        "example": "vpn-pool"
                                    },
                                    "port_Nat": {
                                        "type": "string",
                                        "example": "1194"
                                    },
                                    "port_address": {
                                        "type": "string",
                                        "example": "8729"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil mengkonfigurasi VPN",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "VPN berhasil dikonfigurasi di Mikrotik"
                                        },
                                        "client_ip_range": {
                                            "type": "string",
                                            "example": "172.16.1.1"
                                        },
                                        "address_network": {
                                            "type": "string",
                                            "example": "172.16.1.1"
                                        },
                                        "local_address": {
                                            "type": "string",
                                            "example": "172.16.1.2"
                                        },
                                        "commands": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "/interface ovpn-client add name=client-vpnuser connect-to=45.149.93.122 ..."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Data tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data tidak lengkap atau tidak valid"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan internal"
                    }
                }
            }
        },
        "/configure-masquarade": {
            "post": {
                "tags": [
                    "OpenVPN"
                ],
                "summary": "Tambah NAT Masquerade di Mikrotik",
                "description": "Menambahkan aturan NAT masquerade baru ke Mikrotik menggunakan perintah CLI yang dikirim dari API.",
                "operationId": "2a8fa070d804a15f56d12b81cfd76774",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "command"
                                ],
                                "properties": {
                                    "command": {
                                        "type": "string",
                                        "example": "/ip firewall nat add chain=srcnat out-interface=ovpn-client1 action=masquerade comment=Masquerade_user1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "NAT masquerade berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "NAT masquerade berhasil ditambahkan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Input tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Input does not match any value of interface."
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan internal"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Gunakan Bearer Token untuk autentikasi.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}